# swagger_json.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/policy-config": {
      "post": {
        "operationId": "PolicyConfigController_createPolicy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "策略配置创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      },
      "get": {
        "operationId": "PolicyConfigController_getAllPolicies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "策略配置列表获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      }
    },
    "/api/policy-config/{id}": {
      "put": {
        "operationId": "PolicyConfigController_updatePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "策略配置更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "策略配置不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      },
      "delete": {
        "operationId": "PolicyConfigController_deletePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "策略配置删除成功"
          },
          "404": {
            "description": "策略配置不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      },
      "get": {
        "operationId": "PolicyConfigController_getPolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "策略配置获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "策略配置不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      }
    },
    "/api/policy-config/{id}/enable": {
      "put": {
        "operationId": "PolicyConfigController_enablePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "策略配置启用成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "启用权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      }
    },
    "/api/policy-config/{id}/disable": {
      "put": {
        "operationId": "PolicyConfigController_disablePolicy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "策略配置禁用成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "禁用权限策略配置",
        "tags": [
          "权限策略配置"
        ]
      }
    }
  },
  "info": {
    "title": "CloudCAD API",
    "description": "图纸管理平台API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePolicyDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "策略类型",
            "enum": [
              "TIME",
              "IP",
              "DEVICE"
            ],
            "example": "TIME"
          },
          "name": {
            "type": "string",
            "description": "策略名称",
            "example": "工作时间限制"
          },
          "description": {
            "type": "string",
            "description": "策略描述",
            "example": "仅允许在工作时间 9:00-18:00 访问"
          },
          "config": {
            "type": "object",
            "description": "策略配置",
            "example": {
              "startTime": "09:00",
              "endTime": "18:00",
              "allowedDays": [
                1,
                2,
                3,
                4,
                5
              ]
            }
          },
          "permissions": {
            "type": "array",
            "description": "关联的权限",
            "example": [
              "SYSTEM_USER_DELETE"
            ],
            "items": {
              "type": "string",
              "enum": [
                "SYSTEM_USER_READ",
                "SYSTEM_USER_CREATE",
                "SYSTEM_USER_UPDATE",
                "SYSTEM_USER_DELETE",
                "SYSTEM_ROLE_READ",
                "SYSTEM_ROLE_CREATE",
                "SYSTEM_ROLE_UPDATE",
                "SYSTEM_ROLE_DELETE",
                "SYSTEM_ROLE_PERMISSION_MANAGE",
                "SYSTEM_FONT_READ",
                "SYSTEM_FONT_UPLOAD",
                "SYSTEM_FONT_DELETE",
                "SYSTEM_FONT_DOWNLOAD",
                "SYSTEM_ADMIN",
                "SYSTEM_MONITOR"
              ]
            }
          },
          "enabled": {
            "type": "boolean",
            "description": "是否启用",
            "default": true
          },
          "priority": {
            "type": "number",
            "description": "优先级（数值越大优先级越高）",
            "default": 0
          }
        },
        "required": [
          "type",
          "name",
          "config",
          "permissions"
        ]
      },
      "PolicyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "策略 ID"
          },
          "type": {
            "type": "string",
            "description": "策略类型",
            "enum": [
              "TIME",
              "IP",
              "DEVICE"
            ]
          },
          "name": {
            "type": "string",
            "description": "策略名称"
          },
          "description": {
            "type": "string",
            "description": "策略描述"
          },
          "config": {
            "type": "object",
            "description": "策略配置"
          },
          "permissions": {
            "type": "array",
            "description": "关联的权限",
            "items": {
              "type": "string",
              "enum": [
                "SYSTEM_USER_READ",
                "SYSTEM_USER_CREATE",
                "SYSTEM_USER_UPDATE",
                "SYSTEM_USER_DELETE",
                "SYSTEM_ROLE_READ",
                "SYSTEM_ROLE_CREATE",
                "SYSTEM_ROLE_UPDATE",
                "SYSTEM_ROLE_DELETE",
                "SYSTEM_ROLE_PERMISSION_MANAGE",
                "SYSTEM_FONT_READ",
                "SYSTEM_FONT_UPLOAD",
                "SYSTEM_FONT_DELETE",
                "SYSTEM_FONT_DOWNLOAD",
                "SYSTEM_ADMIN",
                "SYSTEM_MONITOR"
              ]
            }
          },
          "enabled": {
            "type": "boolean",
            "description": "是否启用"
          },
          "priority": {
            "type": "number",
            "description": "优先级"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "config",
          "permissions",
          "enabled",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePolicyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "策略名称",
            "example": "工作时间限制"
          },
          "description": {
            "type": "string",
            "description": "策略描述",
            "example": "仅允许在工作时间 9:00-18:00 访问"
          },
          "config": {
            "type": "object",
            "description": "策略配置",
            "example": {
              "startTime": "09:00",
              "endTime": "18:00",
              "allowedDays": [
                1,
                2,
                3,
                4,
                5
              ]
            }
          },
          "permissions": {
            "type": "array",
            "description": "关联的权限",
            "example": [
              "SYSTEM_USER_DELETE"
            ],
            "items": {
              "type": "string",
              "enum": [
                "SYSTEM_USER_READ",
                "SYSTEM_USER_CREATE",
                "SYSTEM_USER_UPDATE",
                "SYSTEM_USER_DELETE",
                "SYSTEM_ROLE_READ",
                "SYSTEM_ROLE_CREATE",
                "SYSTEM_ROLE_UPDATE",
                "SYSTEM_ROLE_DELETE",
                "SYSTEM_ROLE_PERMISSION_MANAGE",
                "SYSTEM_FONT_READ",
                "SYSTEM_FONT_UPLOAD",
                "SYSTEM_FONT_DELETE",
                "SYSTEM_FONT_DOWNLOAD",
                "SYSTEM_ADMIN",
                "SYSTEM_MONITOR"
              ]
            }
          },
          "enabled": {
            "type": "boolean",
            "description": "是否启用",
            "default": true
          },
          "priority": {
            "type": "number",
            "description": "优先级",
            "default": 0
          },
          "type": {
            "type": "string",
            "description": "策略类型",
            "enum": [
              "TIME",
              "IP",
              "DEVICE"
            ]
          }
        }
      },
...
    }
  }
}